Moves test utils module, fixes spawn_miner args
